Add CurrentUserService

+++ WeTapAPI/Application/Features/Videos/Commands/DeleteVideo/DeleteVideoHandler.cs
@@ -7,12 +7,15 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Application.Features.Videos.Commands.DeleteVideo;
 
-public class DeleteVideoHandler(IGenericRepository<VideoEntity, long> repo)
+public class DeleteVideoHandler(IGenericRepository<VideoEntity, long> repo, ICurrentUserService currentUserService)
     : IRequestHandler<DeleteVideoCommand>
 {
     public async Task Handle(DeleteVideoCommand request, CancellationToken cancellationToken)
     {
         var video = await repo.GetByIdAsync(request.Model.Id);
+
+        if (video!.ChannelId != currentUserService.GetCurrentUserId())
+            throw new Exception("Ви не є власником цього відео");
 
         await repo.DeleteAsync(video.Id);
     }
